Fake News Detection Web App.
This is a web project developed as part of my software engineering graduation. It focuses on Fake News Detection, utilizing three different models: SVM, LSTM, and Random Forest algorithms. The web framework Flask is used to implement the project. The goal is to identify and classify fake news through an interactive web interface.